chore: rename JSX files to .tsx for TypeScript support
