+++ src/app/[locale]/(authUserView)/about/page.tsx
