+++ src/app/[locale]/(authUserView)/blog/[id]/page.tsx
