+++ src/app/[locale]/(authUserView)/blog/[id]/post.tsx
