+++ src/app/[locale]/(authUserView)/blog/page.tsx
