+++ src/app/[locale]/(authUserView)/products/ProductCard.tsx
