+++ src/app/[locale]/(authUserView)/products/[id]/page.tsx
