+++ src/app/[locale]/(authUserView)/products/[id]/product.tsx
