+++ src/app/[locale]/(authUserView)/products/search.tsx
