+++ src/app/[locale]/(authUserView)/profile/page.tsx
