+++ src/app/[locale]/(authUserView)/profile/user-avatar-name.tsx
