+++ src/app/components/welcome/welcome-content.tsx
