+++ src/app/components/welcome/welcome-header.tsx
