+++ src/app/components/welcome/welcome-image.tsx
